[refactor] 데이터타입을 내부/외부를 분리하고, 다운 캐스팅 진행

+++ src/main/java/com/usto/api/ai/forecast/application/ForecastApplication.java
@@ -45,8 +45,8 @@ public class ForecastApplication {
         //도메인 객체에 내용 담기
         Forecast forecast = ForecastMapper.toDomain(
                 usrId,
-                request.conditions().year(),
-                request.conditions().semester(),
+                request.conditions().year().shortValue(),
+                request.conditions().semester().byteValue(),
                 request.conditions().risk_level(),
                 request.prompt(),
                 orgCd,
