+++ src/main/java/com/usto/api/ai/forecast/domain/model/Forecast.java
@@ -12,8 +12,8 @@ import java.util.UUID;
 public class Forecast {
     private final UUID forecastId;
     private final String userId;
-    private final Integer analysisYear;
-    private final Integer semester;
+    private final Short analysisYear;
+    private final Byte semester;
     private final RiskLevel riskLevel;
     private final String message;
     private final String deptCd;
